v0.8.4

+++ norm_scripts/runblast.pl
@@ -380,10 +380,12 @@ foreach my $db2 (keys %DBS2){
     sleep(2);
 }
 if ($type eq "-fq"){
-    `rm $file1 $file2`;
+    if ($se eq "true"){
+	`rm $file1`;
+    }
+    if ($pe eq "true"){
+	`rm $file1 $file2`;
+    }
 }
-my @tqfiles = glob("$tempq*");
-if (@tqfiles > 0){
-    `rm $tempq*`;
-}
+
 print "got here\n";
